langchain-agent/requirements.txt
pydantic
python-dotenv
requests
langchain-core
langchain
langchain_ollama
langchain-google-genai